asap

+++ 0x04-more_functions_nested_loops/100-prime_factor.c
@@ -9,8 +9,8 @@
  */
 int main(void)
 {
-	long long num = 612852475143;
-	long long factor = 2;
+	long num = 612852475143;
+	long factor = 2;
 
 	while (num > 1)
 	{
@@ -24,6 +24,6 @@ int main(void)
 		}
 	}
 
-	printf("%lld\n", factor);
+	printf("%ld\n", factor);
 	return 0;
 }
